src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
azure-common==1.1.28
azure-core==1.30.2
azure-data-tables==12.5.0
azure-identity==1.17.1
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-subscription==3.1.1
azure-monitor-opentelemetry-exporter==1.0.0b28
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
fastapi==0.112.2
fixedint==0.1.6
freezegun==1.5.1
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
numpy==2.1.0
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
portalocker==2.10.1
psutil==5.9.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.1
